+++ source/production/F0.Common/Mathematics/Comparable.cs
@@ -27,5 +27,23 @@ namespace F0.Mathematics
 
 			return value;
 		}
+
+		public static T Max<T>(T x, T y)
+			where T : IComparable<T>
+		{
+			_ = x ?? throw new ArgumentNullException(nameof(x));
+			_ = y ?? throw new ArgumentNullException(nameof(y));
+
+			return x.CompareTo(y) >= 0 ? x : y;
+		}
+
+		public static T Min<T>(T x, T y)
+			where T : IComparable<T>
+		{
+			_ = x ?? throw new ArgumentNullException(nameof(x));
+			_ = y ?? throw new ArgumentNullException(nameof(y));
+
+			return x.CompareTo(y) <= 0 ? x : y;
+		}
 	}
 }